Creando modelo ContratistaContrato

+++ database/migrations/2022_11_22_145715_create_contratos_table.php
@@ -26,7 +26,7 @@ class CreateContratosTable extends Migration
             $table->integer('unspsc');
             $table->string('unspsc_adicionales')->comment('Unspsc adicionales separados por coma');
             $table->integer('numero_documentos');
-            $table->integer('valor')->nullable();
+            $table->bigInteger('valor')->nullable();
             $table->string('valor_texto')->nullable();
             $table->date('fecha_actualizacion_estado')->comment('Esta fecha solo debe actualizarse cuando cambie el estado_agrupado');
             $table->date('fecha_last_update_seguimiento')->comment('Fecha utilizada para manejar el seguimiento, con esto sabemos la ultima vez que se miro el contrato desde seguimiento');
